+++ swiftide-test-utils/src/lib.rs
@@ -1,0 +1,8 @@
+#[cfg(feature = "test-utils")]
+mod test_utils;
+
+
+#[cfg(feature = "test-utils")]
+pub use test_utils::*;
+
+
