+++ swiftide/tests/indexing_pipeline.rs
@@ -1,19 +1,15 @@
+#![cfg(test)]
 //! This module contains tests for the indexing pipeline in the Swiftide project.
 //! The tests validate the functionality of the pipeline, ensuring it processes data correctly
 //! from a temporary file, simulates API responses, and stores data accurately in the Qdrant vector database.
 
 use qdrant_client::qdrant::vectors::VectorsOptions;
 use qdrant_client::qdrant::{SearchPointsBuilder, Value};
-use serde_json::json;
-use swiftide::indexing;
 use swiftide::indexing::*;
 use swiftide::integrations;
+use swiftide_test_utils::*;
 use temp_dir::TempDir;
-use testcontainers::core::wait::HttpWaitStrategy;
-use testcontainers::runners::AsyncRunner;
-use testcontainers::{ContainerAsync, GenericImage};
-use wiremock::matchers::{method, path};
-use wiremock::{Mock, MockServer, ResponseTemplate};
+use wiremock::MockServer;
 
 /// Tests the indexing pipeline without any mocks.
 ///
@@ -158,7 +154,7 @@ async fn test_named_vectors() {
 
     let result =
         Pipeline::from_loader(loaders::FileLoader::new(tempdir.path()).with_extensions(&["rs"]))
-            .with_embed_mode(indexing::EmbedMode::PerField)
+            .with_embed_mode(EmbedMode::PerField)
             .then_chunk(transformers::ChunkCode::try_for_language("rust").unwrap())
             .then(transformers::MetadataQACode::new(openai_client.clone()))
             .filter_cached(integrations::redis::Redis::try_from_url(&redis_url, "prefix").unwrap())
@@ -230,121 +226,4 @@ async fn test_named_vectors() {
     assert!(vectors.contains_key(
         &EmbeddedField::Metadata(transformers::metadata_qa_code::NAME.into()).to_string()
     ));
-}
-
-/// Setup `OpenAI` client with the mock server
-fn openai_client(
-    mock_server_uri: &str,
-    embed_model: &str,
-    prompt_model: &str,
-) -> integrations::openai::OpenAI {
-    let config = async_openai::config::OpenAIConfig::new().with_api_base(mock_server_uri);
-    let async_openai = async_openai::Client::with_config(config);
-    integrations::openai::OpenAI::builder()
-        .client(async_openai)
-        .default_options(
-            integrations::openai::Options::builder()
-                .embed_model(embed_model)
-                .prompt_model(prompt_model)
-                .build()
-                .unwrap(),
-        )
-        .build()
-        .expect("Can create OpenAI client.")
-}
-
-/// Setup Qdrant container.
-/// Returns container server and `server_url`.
-async fn start_qdrant() -> (ContainerAsync<GenericImage>, String) {
-    let qdrant = testcontainers::GenericImage::new("qdrant/qdrant", "v1.9.2")
-        .with_exposed_port(6334.into())
-        .with_exposed_port(6333.into())
-        .with_wait_for(testcontainers::core::WaitFor::http(
-            HttpWaitStrategy::new("/readyz")
-                .with_port(6333.into())
-                .with_expected_status_code(200_u16),
-        ))
-        .start()
-        .await
-        .expect("Qdrant started");
-    let qdrant_url = format!(
-        "http://127.0.0.1:{port}",
-        port = qdrant.get_host_port_ipv4(6334).await.unwrap()
-    );
-    (qdrant, qdrant_url)
-}
-
-/// Setup Redis container for caching in the test.
-/// Returns container server and `server_url`.
-async fn start_redis() -> (ContainerAsync<GenericImage>, String) {
-    let redis = testcontainers::GenericImage::new("redis", "7.2.4")
-        .with_exposed_port(6379.into())
-        .with_wait_for(testcontainers::core::WaitFor::message_on_stdout(
-            "Ready to accept connections",
-        ))
-        .start()
-        .await
-        .expect("Redis started");
-    let redis_url = format!(
-        "redis://{host}:{port}",
-        host = redis.get_host().await.unwrap(),
-        port = redis.get_host_port_ipv4(6379).await.unwrap()
-    );
-    (redis, redis_url)
-}
-
-/// Mock embeddings creation endpoint.
-/// `embeddings_count` controls number of returned embedding vectors.
-async fn mock_embeddings(mock_server: &MockServer, embeddings_count: u8) {
-    let data = (0..embeddings_count)
-        .map(|i| {
-            json!( {
-              "object": "embedding",
-              "embedding": vec![0; 1536],
-              "index": i
-            })
-        })
-        .collect::<Vec<serde_json::Value>>();
-    let data: serde_json::Value = serde_json::Value::Array(data);
-    Mock::given(method("POST"))
-        .and(path("/embeddings"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-          "object": "list",
-          "data": data,
-          "model": "text-embedding-ada-002",
-          "usage": {
-            "prompt_tokens": 8,
-            "total_tokens": 8
-        }
-        })))
-        .mount(mock_server)
-        .await;
-}
-
-async fn mock_chat_completions(mock_server: &MockServer) {
-    Mock::given(method("POST"))
-        .and(path("/chat/completions"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-            "id": "chatcmpl-123",
-            "object": "chat.completion",
-            "created": 1_677_652_288,
-            "model": "gpt-3.5-turbo-0125",
-            "system_fingerprint": "fp_44709d6fcb",
-            "choices": [{
-              "index": 0,
-              "message": {
-                "role": "assistant",
-                "content": "\n\nHello there, how may I assist you today?",
-              },
-              "logprobs": null,
-              "finish_reason": "stop"
-            }],
-            "usage": {
-              "prompt_tokens": 9,
-              "completion_tokens": 12,
-              "total_tokens": 21
-            }
-        })))
-        .mount(mock_server)
-        .await;
 }
